LBSD-2200 Refactored default theme to allow multiple levels of nesting dependency - Splitted tasks into separate files in order to make it easy to maintain - Reworked a few pieces to achieve new structure to handle multiple levels
