+++ server/liveblog/themes/themes_assets/default/package.json
@@ -1,6 +1,6 @@
 {
   "name": "liveblog-default-theme",
-  "version": "3.3.93",
+  "version": "4.0.0-beta.1",
   "description": "Liveblog 3 SEO theme",
   "author": "Paul Solbach <psolbach@dpa-newslab.com>, Massimo Scamarcia <massimo.scamarcia@sourcefabric.org>",
   "license": "AGPL-3.0",
@@ -43,9 +43,11 @@
     "liveblog-shared-tools": "eos87/liveblog-shared-tools",
     "lory.js": "^2.3.4",
     "moment": "^2.18.1",
+    "npm-exists": "^1.0.1",
     "nunjucks": "~3.0.1",
     "nunjucks-date-filter": "^0.1.1",
     "promise-polyfill": "^6.0.2",
+    "spawn-npm-install": "^1.2.0",
     "through2": "^2.0.3",
     "vinyl-buffer": "^1.0.0",
     "vinyl-source-stream": "^1.1.0"
