+++ server/liveblog/themes/themes_assets/default/tasks/constants.js
@@ -1,0 +1,6 @@
+module.exports = {
+  TEMPLATES_SUFFIX:  'precomp-templates.js',
+  PRE_BUNDLE_DIR: 'dist/prebundle/',
+  BUILD_HTML: './index.html',
+  SYSTEM_THEMES: ['amp', 'default', 'simple'] // extendable system themes
+}
