+++ server/liveblog/themes/themes_assets/default/theme.json
@@ -1,7 +1,7 @@
 {
     "label": "Liveblog 3 Default SEO Theme",
     "name": "default",
-    "version": "3.3.93",
+    "version": "4.0.0-beta.1",
     "asyncTheme": true,
     "seoTheme": true,
     "author": "Sourcefabric <contact@sourcefabric.org>",
@@ -9,7 +9,8 @@
         "Paul Solbach <psolbach@dpa-newslab.com>",
         "Massimo Scamarcia <massimo.scamarcia@sourcefabric.org>",
         "Löic Nogues <loic.nogues@sourcefabric.org>",
-        "Aleksandar Backo Jelicic <aleksandar.jelicic@sourcefabric.org>"
+        "Aleksandar Backo Jelicic <aleksandar.jelicic@sourcefabric.org>",
+        "Helmy Giacoman <helmy.giacoman@sourcefabric.org>"
     ],
     "repository": {
         "type" : "git",
